requirements.txt
absl-py==1.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
bleach==4.1.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.4
colorama==0.4.4
cryptography==36.0.1
cycler==0.11.0
Cython==0.29.26
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
entrypoints==0.3
filelock==3.4.2
Flask==2.0.3
flatbuffers==2.0
fonttools==4.28.5
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.19.1
grpcio==1.43.0
h11==0.12.0
h5py==3.6.0
idna==3.3
imageio==2.13.5
imutils==0.5.4
ipykernel==6.6.1
ipython==7.31.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.1.0
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.3.3
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==13.0.0
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.9
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
notebook==6.4.6
numpy==1.22.0
oauthlib==3.2.0
opencv-python==4.5.5.62
opt-einsum==3.3.0
outcome==1.1.0
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.4.1
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydotplus==2.0.2
Pygments==2.11.2
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyperclip==1.8.2
pyrsistent==0.18.0
pytesseract==0.3.8
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2021.3
PyWavelets==1.2.0
pywin32==303
pywinpty==1.1.6
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
selenium==4.1.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
subprocess.run==0.0.8
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
terminado==0.12.1
tesseract==0.1.3
testpath==0.5.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.0.0
tifffile==2021.11.2
tornado==6.1
traitlets==5.1.1
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.1
urllib3==1.26.8
virtualenv==20.13.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
word2number==1.1
wrapt==1.13.3
wsproto==1.0.0
